pyproject.toml
[tool.poetry]
name = "homelab306"
version = "0.1.0"
description = ""
authors = ["Emma Knospe <emma@emmaknospe.dev>"]
readme = "README.md"
packages = [
    { include = "cli", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
paramiko = "^3.5.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
homelab = "cli.main:cli"
